gitserver: Organize server code better
This PR is purely mechanical. It groups responsibilities a bit better in separate files vs the 1700 loc server megafile.

The logic itself is pretty much untouched.
