+++ cmd/gitserver/internal/server.go
@@ -2,12 +2,7 @@
 package internal
 
 import (
-	"bufio"
-	"bytes"
-	"container/list"
 	"context"
-	"fmt"
-	"io"
 	"net/http"
 	"os"
 	"os/exec"
@@ -20,10 +15,7 @@ import (
 
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
-	"go.opentelemetry.io/otel/attribute"
-	"golang.org/x/sync/errgroup"
 	"golang.org/x/sync/semaphore"
-	"golang.org/x/time/rate"
 
 	"github.com/sourcegraph/log"
 
@@ -40,19 +32,14 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/env"
-	"github.com/sourcegraph/sourcegraph/internal/featureflag"
 	"github.com/sourcegraph/sourcegraph/internal/fileutil"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver"
-	"github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver/protocol"
-	"github.com/sourcegraph/sourcegraph/internal/goroutine"
-	"github.com/sourcegraph/sourcegraph/internal/honey"
 	"github.com/sourcegraph/sourcegraph/internal/lazyregexp"
 	"github.com/sourcegraph/sourcegraph/internal/limiter"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
 	"github.com/sourcegraph/sourcegraph/internal/ratelimit"
 	"github.com/sourcegraph/sourcegraph/internal/trace"
-	"github.com/sourcegraph/sourcegraph/internal/types"
 	"github.com/sourcegraph/sourcegraph/internal/vcs"
 	"github.com/sourcegraph/sourcegraph/internal/wrexec"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
@@ -83,34 +70,6 @@ func debounce(name api.RepoName, since time.Duration) bool {
 
 func init() {
 	traceLogs, _ = strconv.ParseBool(env.Get("SRC_GITSERVER_TRACE", "false", "Toggles trace logging to stderr"))
-}
-
-// cloneJob abstracts away a repo and necessary metadata to clone it. In the future it may be
-// possible to simplify this, but to do that, doClone will need to do a lot less than it does at the
-// moment.
-type cloneJob struct {
-	repo   api.RepoName
-	dir    common.GitDir
-	syncer vcssyncer.VCSSyncer
-
-	// TODO: cloneJobConsumer should acquire a new lock. We are trying to keep the changes simple
-	// for the time being. When we start using the new approach of using long lived goroutines for
-	// cloning we will refactor doClone to acquire a new lock.
-	lock RepositoryLock
-
-	remoteURL *vcs.URL
-	options   CloneOptions
-}
-
-// cloneTask is a thin wrapper around a cloneJob to associate the doneFunc with each job.
-type cloneTask struct {
-	*cloneJob
-	done func() time.Duration
-}
-
-// NewCloneQueue initializes a new cloneQueue.
-func NewCloneQueue(obctx *observation.Context, jobs *list.List) *common.Queue[*cloneJob] {
-	return common.NewQueue[*cloneJob](obctx, "clone-queue", jobs)
 }
 
 // Server is a gitserver server.
@@ -201,25 +160,6 @@ type locks struct {
 	mu   *sync.Mutex // prevents updates running in parallel
 }
 
-// shortGitCommandSlow returns the threshold for regarding an git command as
-// slow. Some commands such as "git archive" are inherently slower than "git
-// rev-parse", so this will return an appropriate threshold given the command.
-func shortGitCommandSlow(args []string) time.Duration {
-	if len(args) < 1 {
-		return time.Second
-	}
-	switch args[0] {
-	case "archive":
-		return 1 * time.Minute
-
-	case "blame", "ls-tree", "log", "show":
-		return 5 * time.Second
-
-	default:
-		return 2500 * time.Millisecond
-	}
-}
-
 // Handler returns the http.Handler that should be used to serve requests.
 func (s *Server) Handler() http.Handler {
 	s.ctx, s.cancel = context.WithCancel(context.Background())
@@ -270,110 +210,6 @@ func (s *Server) Handler() http.Handler {
 
 func addrForRepo(ctx context.Context, repoName api.RepoName, gitServerAddrs gitserver.GitserverAddresses) string {
 	return gitServerAddrs.AddrForRepo(ctx, filepath.Base(os.Args[0]), repoName)
-}
-
-// NewClonePipeline creates a new pipeline that clones repos asynchronously. It
-// creates a producer-consumer pipeline that handles clone requests asychronously.
-func (s *Server) NewClonePipeline(logger log.Logger, cloneQueue *common.Queue[*cloneJob]) goroutine.BackgroundRoutine {
-	return &clonePipelineRoutine{
-		tasks:  make(chan *cloneTask),
-		logger: logger,
-		s:      s,
-		queue:  cloneQueue,
-	}
-}
-
-type clonePipelineRoutine struct {
-	logger log.Logger
-
-	tasks chan *cloneTask
-	// TODO: Get rid of this dependency.
-	s      *Server
-	queue  *common.Queue[*cloneJob]
-	cancel context.CancelFunc
-}
-
-func (p *clonePipelineRoutine) Start() {
-	ctx, cancel := context.WithCancel(context.Background())
-	p.cancel = cancel
-	// Start a go routine for each the producer and the consumer.
-	go p.cloneJobConsumer(ctx, p.tasks)
-	go p.cloneJobProducer(ctx, p.tasks)
-}
-
-func (p *clonePipelineRoutine) Stop() {
-	if p.cancel != nil {
-		p.cancel()
-	}
-}
-
-func (p *clonePipelineRoutine) cloneJobProducer(ctx context.Context, tasks chan<- *cloneTask) {
-	defer close(tasks)
-
-	for {
-		// Acquire the cond mutex lock and wait for a signal if the queue is empty.
-		p.queue.Mutex.Lock()
-		if p.queue.Empty() {
-			// TODO: This should only wait if ctx is not canceled.
-			p.queue.Cond.Wait()
-		}
-
-		// The queue is not empty and we have a job to process! But don't forget to unlock the cond
-		// mutex here as we don't need to hold the lock beyond this point for now.
-		p.queue.Mutex.Unlock()
-
-		// Keep popping from the queue until the queue is empty again, in which case we start all
-		// over again from the top.
-		for {
-			job, doneFunc := p.queue.Pop()
-			if job == nil {
-				break
-			}
-
-			select {
-			case tasks <- &cloneTask{
-				cloneJob: *job,
-				done:     doneFunc,
-			}:
-			case <-ctx.Done():
-				p.logger.Error("cloneJobProducer", log.Error(ctx.Err()))
-				return
-			}
-		}
-	}
-}
-
-func (p *clonePipelineRoutine) cloneJobConsumer(ctx context.Context, tasks <-chan *cloneTask) {
-	logger := p.s.Logger.Scoped("cloneJobConsumer")
-
-	for task := range tasks {
-		logger := logger.With(log.String("job.repo", string(task.repo)))
-
-		select {
-		case <-ctx.Done():
-			logger.Error("context done", log.Error(ctx.Err()))
-			return
-		default:
-		}
-
-		ctx, cancel, err := p.s.acquireCloneLimiter(ctx)
-		if err != nil {
-			logger.Error("acquireCloneLimiter", log.Error(err))
-			continue
-		}
-
-		go func(task *cloneTask) {
-			defer cancel()
-
-			err := p.s.doClone(ctx, task.repo, task.dir, task.syncer, task.lock, task.remoteURL, task.options)
-			if err != nil {
-				logger.Error("failed to clone repo", log.Error(err))
-			}
-			// Use a different context in case we failed because the original context failed.
-			p.s.setLastErrorNonFatal(p.s.ctx, task.repo, err)
-			_ = task.done()
-		}(task)
-	}
 }
 
 // repoCloned checks if dir or `${dir}/.git` is a valid GIT_DIR.
@@ -429,6 +265,17 @@ func (s *Server) getRemoteURL(ctx context.Context, name api.RepoName) (*vcs.URL,
 	return vcs.ParseURL(remoteURL)
 }
 
+var (
+	pendingClones = promauto.NewGauge(prometheus.GaugeOpts{
+		Name: "src_gitserver_clone_queue",
+		Help: "number of repos waiting to be cloned.",
+	})
+	lsRemoteQueue = promauto.NewGauge(prometheus.GaugeOpts{
+		Name: "src_gitserver_lsremote_queue",
+		Help: "number of repos waiting to check existence on remote code host (git ls-remote).",
+	})
+)
+
 // acquireCloneLimiter() acquires a cancellable context associated with the
 // clone limiter.
 func (s *Server) acquireCloneLimiter(ctx context.Context) (context.Context, context.CancelFunc, error) {
@@ -470,166 +317,6 @@ func (s *Server) isRepoCloneable(ctx context.Context, repo api.RepoName) (protoc
 	return resp, nil
 }
 
-func (s *Server) repoUpdate(req *protocol.RepoUpdateRequest) protocol.RepoUpdateResponse {
-	logger := s.Logger.Scoped("handleRepoUpdate")
-	var resp protocol.RepoUpdateResponse
-	req.Repo = protocol.NormalizeRepo(req.Repo)
-	dir := gitserverfs.RepoDirFromName(s.ReposDir, req.Repo)
-
-	// despite the existence of a context on the request, we don't want to
-	// cancel the git commands partway through if the request terminates.
-	ctx, cancel1 := s.serverContext()
-	defer cancel1()
-	ctx, cancel2 := context.WithTimeout(ctx, conf.GitLongCommandTimeout())
-	defer cancel2()
-
-	if !repoCloned(dir) && !s.skipCloneForTests {
-		_, err := s.CloneRepo(ctx, req.Repo, CloneOptions{Block: true})
-		if err != nil {
-			logger.Warn("error cloning repo", log.String("repo", string(req.Repo)), log.Error(err))
-			resp.Error = err.Error()
-		}
-		return resp
-	}
-
-	var statusErr, updateErr error
-
-	if debounce(req.Repo, req.Since) {
-		updateErr = s.doRepoUpdate(ctx, req.Repo, "")
-	}
-
-	// attempts to acquire these values are not contingent on the success of
-	// the update.
-	lastFetched, err := repoLastFetched(dir)
-	if err != nil {
-		statusErr = err
-	} else {
-		resp.LastFetched = &lastFetched
-	}
-	lastChanged, err := repoLastChanged(dir)
-	if err != nil {
-		statusErr = err
-	} else {
-		resp.LastChanged = &lastChanged
-	}
-	if statusErr != nil {
-		logger.Error("failed to get status of repo", log.String("repo", string(req.Repo)), log.Error(statusErr))
-		// report this error in-band, but still produce a valid response with the
-		// other information.
-		resp.Error = statusErr.Error()
-	}
-	// If an error occurred during update, report it but don't actually make
-	// it into an http error; we want the client to get the information cleanly.
-	// An update error "wins" over a status error.
-	if updateErr != nil {
-		resp.Error = updateErr.Error()
-	} else {
-		s.Perforce.EnqueueChangelistMappingJob(perforce.NewChangelistMappingJob(req.Repo, dir))
-	}
-
-	return resp
-}
-
-func (s *Server) performGitLogCommand(ctx context.Context, repoCommit api.RepoCommit, format string) (output string, isRepoCloned bool, err error) {
-	ctx, _, endObservation := s.operations.batchLogSingle.With(ctx, &err, observation.Args{
-		Attrs: append(
-			[]attribute.KeyValue{
-				attribute.String("format", format),
-			},
-			repoCommit.Attrs()...,
-		),
-	})
-	defer func() {
-		endObservation(1, observation.Args{Attrs: []attribute.KeyValue{
-			attribute.Bool("isRepoCloned", isRepoCloned),
-		}})
-	}()
-
-	dir := gitserverfs.RepoDirFromName(s.ReposDir, repoCommit.Repo)
-	if !repoCloned(dir) {
-		return "", false, nil
-	}
-
-	var buf bytes.Buffer
-
-	commitId := string(repoCommit.CommitID)
-	// make sure CommitID is not an arg
-	if commitId[0] == '-' {
-		return "", true, errors.New("commit ID starting with - is not allowed")
-	}
-
-	cmd := s.RecordingCommandFactory.Command(ctx, s.Logger, string(repoCommit.Repo), "git", "log", "-n", "1", "--name-only", format, commitId)
-	dir.Set(cmd.Unwrap())
-	cmd.Unwrap().Stdout = &buf
-
-	if _, err := executil.RunCommand(ctx, cmd); err != nil {
-		return "", true, err
-	}
-
-	return buf.String(), true, nil
-}
-
-func (s *Server) batchGitLogInstrumentedHandler(ctx context.Context, req protocol.BatchLogRequest) (resp protocol.BatchLogResponse, err error) {
-	ctx, _, endObservation := s.operations.batchLog.With(ctx, &err, observation.Args{})
-	defer func() {
-		endObservation(1, observation.Args{Attrs: []attribute.KeyValue{
-			attribute.String("results", fmt.Sprintf("%+v", resp.Results)),
-		}})
-	}()
-
-	// Perform requests in each repository in the input batch. We perform these commands
-	// concurrently, but only allow for so many commands to be in-flight at a time so that
-	// we don't overwhelm a shard with either a large request or too many concurrent batch
-	// requests.
-
-	g, ctx := errgroup.WithContext(ctx)
-	results := make([]protocol.BatchLogResult, len(req.RepoCommits))
-
-	if s.GlobalBatchLogSemaphore == nil {
-		return protocol.BatchLogResponse{}, errors.New("s.GlobalBatchLogSemaphore not initialized")
-	}
-
-	for i, repoCommit := range req.RepoCommits {
-		// Avoid capture of loop variables
-		i, repoCommit := i, repoCommit
-
-		start := time.Now()
-		if err := s.GlobalBatchLogSemaphore.Acquire(ctx, 1); err != nil {
-			return resp, err
-		}
-		s.operations.batchLogSemaphoreWait.Observe(time.Since(start).Seconds())
-
-		g.Go(func() error {
-			defer s.GlobalBatchLogSemaphore.Release(1)
-
-			output, isRepoCloned, gitLogErr := s.performGitLogCommand(ctx, repoCommit, req.Format)
-			if gitLogErr == nil && !isRepoCloned {
-				gitLogErr = errors.Newf("repo not found")
-			}
-			var errMessage string
-			if gitLogErr != nil {
-				errMessage = gitLogErr.Error()
-			}
-
-			// Concurrently write results to shared slice. This slice is already properly
-			// sized, and each goroutine writes to a unique index exactly once. There should
-			// be no data race conditions possible here.
-
-			results[i] = protocol.BatchLogResult{
-				RepoCommit:    repoCommit,
-				CommandOutput: output,
-				CommandError:  errMessage,
-			}
-			return nil
-		})
-	}
-
-	if err = g.Wait(); err != nil {
-		return
-	}
-	return protocol.BatchLogResponse{Results: results}, nil
-}
-
 // ensureOperations returns the non-nil operations value supplied to this server
 // via RegisterMetrics (when constructed as part of the gitserver binary), or
 // constructs and memoizes a no-op operations value (for use in tests).
@@ -639,208 +326,6 @@ func (s *Server) ensureOperations() *operations {
 	}
 
 	return s.operations
-}
-
-var blockedCommandExecutedCounter = promauto.NewCounter(prometheus.CounterOpts{
-	Name: "src_gitserver_exec_blocked_command_received",
-	Help: "Incremented each time a command not in the allowlist for gitserver is executed",
-})
-
-var ErrInvalidCommand = errors.New("invalid command")
-
-type NotFoundError struct {
-	Payload *protocol.NotFoundPayload
-}
-
-func (e *NotFoundError) Error() string { return "not found" }
-
-type execStatus struct {
-	ExitStatus int
-	Stderr     string
-	Err        error
-}
-
-// TODO: eseliger
-// exec runs a git command. After the first write to w, it must not return an error.
-// TODO(@camdencheek): once gRPC is the only consumer of this, do everything with errors
-// because gRPC can handle trailing errors on a stream.
-func (s *Server) exec(ctx context.Context, logger log.Logger, req *protocol.ExecRequest, userAgent string, w io.Writer) (execStatus, error) {
-	// 🚨 SECURITY: Ensure that only commands in the allowed list are executed.
-	// See https://github.com/sourcegraph/security-issues/issues/213.
-
-	repoPath := string(protocol.NormalizeRepo(req.Repo))
-	repoDir := filepath.Join(s.ReposDir, filepath.FromSlash(repoPath))
-
-	if !gitdomain.IsAllowedGitCmd(logger, req.Args, repoDir) {
-		blockedCommandExecutedCounter.Inc()
-		return execStatus{}, ErrInvalidCommand
-	}
-
-	if !req.NoTimeout {
-		var cancel context.CancelFunc
-		ctx, cancel = context.WithTimeout(ctx, executil.ShortGitCommandTimeout(req.Args))
-		defer cancel()
-	}
-
-	start := time.Now()
-	var cmdStart time.Time // set once we have ensured commit
-	exitStatus := executil.UnsetExitStatus
-	var stdoutN, stderrN int64
-	var status string
-	var execErr error
-	ensureRevisionStatus := "noop"
-
-	req.Repo = protocol.NormalizeRepo(req.Repo)
-	repoName := req.Repo
-
-	// Instrumentation
-	{
-		cmd := ""
-		if len(req.Args) > 0 {
-			cmd = req.Args[0]
-		}
-		args := strings.Join(req.Args, " ")
-
-		var tr trace.Trace
-		tr, ctx = trace.New(ctx, "exec."+cmd, repoName.Attr())
-		tr.SetAttributes(
-			attribute.String("args", args),
-			attribute.String("ensure_revision", req.EnsureRevision),
-		)
-		logger = logger.WithTrace(trace.Context(ctx))
-
-		execRunning.WithLabelValues(cmd).Inc()
-		defer func() {
-			tr.AddEvent(
-				"done",
-				attribute.String("status", status),
-				attribute.Int64("stdout", stdoutN),
-				attribute.Int64("stderr", stderrN),
-				attribute.String("ensure_revision_status", ensureRevisionStatus),
-			)
-			tr.SetError(execErr)
-			tr.End()
-
-			duration := time.Since(start)
-			execRunning.WithLabelValues(cmd).Dec()
-			execDuration.WithLabelValues(cmd, status).Observe(duration.Seconds())
-
-			var cmdDuration time.Duration
-			var fetchDuration time.Duration
-			if !cmdStart.IsZero() {
-				cmdDuration = time.Since(cmdStart)
-				fetchDuration = cmdStart.Sub(start)
-			}
-
-			isSlow := cmdDuration > shortGitCommandSlow(req.Args)
-			isSlowFetch := fetchDuration > 10*time.Second
-			if honey.Enabled() || traceLogs || isSlow || isSlowFetch {
-				act := actor.FromContext(ctx)
-				ev := honey.NewEvent("gitserver-exec")
-				ev.SetSampleRate(honeySampleRate(cmd, act))
-				ev.AddField("repo", repoName)
-				ev.AddField("cmd", cmd)
-				ev.AddField("args", args)
-				ev.AddField("actor", act.UIDString())
-				ev.AddField("ensure_revision", req.EnsureRevision)
-				ev.AddField("ensure_revision_status", ensureRevisionStatus)
-				ev.AddField("client", userAgent)
-				ev.AddField("duration_ms", duration.Milliseconds())
-				ev.AddField("stdin_size", len(req.Stdin))
-				ev.AddField("stdout_size", stdoutN)
-				ev.AddField("stderr_size", stderrN)
-				ev.AddField("exit_status", exitStatus)
-				ev.AddField("status", status)
-				if execErr != nil {
-					ev.AddField("error", execErr.Error())
-				}
-				if !cmdStart.IsZero() {
-					ev.AddField("cmd_duration_ms", cmdDuration.Milliseconds())
-					ev.AddField("fetch_duration_ms", fetchDuration.Milliseconds())
-				}
-
-				if traceID := trace.ID(ctx); traceID != "" {
-					ev.AddField("traceID", traceID)
-					ev.AddField("trace", trace.URL(traceID, conf.DefaultClient()))
-				}
-
-				if honey.Enabled() {
-					_ = ev.Send()
-				}
-
-				if traceLogs {
-					logger.Debug("TRACE gitserver exec", log.Object("ev.Fields", mapToLoggerField(ev.Fields())...))
-				}
-				if isSlow {
-					logger.Warn("Long exec request", log.Object("ev.Fields", mapToLoggerField(ev.Fields())...))
-				}
-				if isSlowFetch {
-					logger.Warn("Slow fetch/clone for exec request", log.Object("ev.Fields", mapToLoggerField(ev.Fields())...))
-				}
-			}
-		}()
-	}
-
-	if notFoundPayload, cloned := s.maybeStartClone(ctx, logger, repoName); !cloned {
-		if notFoundPayload.CloneInProgress {
-			status = "clone-in-progress"
-		} else {
-			status = "repo-not-found"
-		}
-
-		return execStatus{}, &NotFoundError{notFoundPayload}
-	}
-
-	dir := gitserverfs.RepoDirFromName(s.ReposDir, repoName)
-	if s.ensureRevision(ctx, repoName, req.EnsureRevision, dir) {
-		ensureRevisionStatus = "fetched"
-	}
-
-	// Special-case `git rev-parse HEAD` requests. These are invoked by search queries for every repo in scope.
-	// For searches over large repo sets (> 1k), this leads to too many child process execs, which can lead
-	// to a persistent failure mode where every exec takes > 10s, which is disastrous for gitserver performance.
-	if len(req.Args) == 2 && req.Args[0] == "rev-parse" && req.Args[1] == "HEAD" {
-		if resolved, err := git.QuickRevParseHead(dir); err == nil && gitdomain.IsAbsoluteRevision(resolved) {
-			_, _ = w.Write([]byte(resolved))
-			return execStatus{}, nil
-		}
-	}
-
-	// Special-case `git symbolic-ref HEAD` requests. These are invoked by resolvers determining the default branch of a repo.
-	// For searches over large repo sets (> 1k), this leads to too many child process execs, which can lead
-	// to a persistent failure mode where every exec takes > 10s, which is disastrous for gitserver performance.
-	if len(req.Args) == 2 && req.Args[0] == "symbolic-ref" && req.Args[1] == "HEAD" {
-		if resolved, err := git.QuickSymbolicRefHead(dir); err == nil {
-			_, _ = w.Write([]byte(resolved))
-			return execStatus{}, nil
-		}
-	}
-
-	var stderrBuf bytes.Buffer
-	stdoutW := &writeCounter{w: w}
-	stderrW := &writeCounter{w: &limitWriter{W: &stderrBuf, N: 1024}}
-
-	cmdStart = time.Now()
-	cmd := s.RecordingCommandFactory.Command(ctx, s.Logger, string(repoName), "git", req.Args...)
-	dir.Set(cmd.Unwrap())
-	cmd.Unwrap().Stdout = stdoutW
-	cmd.Unwrap().Stderr = stderrW
-	cmd.Unwrap().Stdin = bytes.NewReader(req.Stdin)
-
-	exitStatus, execErr = executil.RunCommand(ctx, cmd)
-
-	status = strconv.Itoa(exitStatus)
-	stdoutN = stdoutW.n
-	stderrN = stderrW.n
-
-	stderr := stderrBuf.String()
-	s.logIfCorrupt(ctx, repoName, dir, stderr)
-
-	return execStatus{
-		Err:        execErr,
-		Stderr:     stderr,
-		ExitStatus: exitStatus,
-	}, nil
 }
 
 func setLastFetched(ctx context.Context, db database.DB, shardID string, dir common.GitDir, name api.RepoName) error {
@@ -862,14 +347,14 @@ func setLastFetched(ctx context.Context, db database.DB, shardID string, dir com
 }
 
 // setLastErrorNonFatal will set the last_error column for the repo in the gitserver table.
-func (s *Server) setLastErrorNonFatal(ctx context.Context, name api.RepoName, err error) {
+func setLastErrorNonFatal(ctx context.Context, logger log.Logger, db database.DB, hostname string, name api.RepoName, err error) {
 	var errString string
 	if err != nil {
 		errString = err.Error()
 	}
 
-	if err := s.DB.GitserverRepos().SetLastError(ctx, name, errString, s.Hostname); err != nil {
-		s.Logger.Warn("Setting last error in DB", log.Error(err))
+	if err := db.GitserverRepos().SetLastError(ctx, name, errString, hostname); err != nil {
+		logger.Warn("Setting last error in DB", log.Error(err))
 	}
 }
 
@@ -926,293 +411,6 @@ func stdErrIndicatesCorruption(stderr string) bool {
 // testRepoCorrupter is used by tests to disrupt a cloned repository (e.g. deleting
 // HEAD, zeroing it out, etc.)
 var testRepoCorrupter func(ctx context.Context, tmpDir common.GitDir)
-
-// cloneOptions specify optional behaviour for the cloneRepo function.
-type CloneOptions struct {
-	// Block will wait for the clone to finish before returning. If the clone
-	// fails, the error will be returned. The passed in context is
-	// respected. When not blocking the clone is done with a server background
-	// context.
-	Block bool
-
-	// Overwrite will overwrite the existing clone.
-	Overwrite bool
-}
-
-// CloneRepo performs a clone operation for the given repository. It is
-// non-blocking by default.
-func (s *Server) CloneRepo(ctx context.Context, repo api.RepoName, opts CloneOptions) (cloneProgress string, err error) {
-	if isAlwaysCloningTest(repo) {
-		return "This will never finish cloning", nil
-	}
-
-	dir := gitserverfs.RepoDirFromName(s.ReposDir, repo)
-
-	// PERF: Before doing the network request to check if isCloneable, lets
-	// ensure we are not already cloning.
-	if progress, cloneInProgress := s.Locker.Status(dir); cloneInProgress {
-		return progress, nil
-	}
-
-	// We always want to store whether there was an error cloning the repo, but only
-	// after we checked if a clone is already in progress, otherwise we would race with
-	// the actual running clone for the DB state of last_error.
-	defer func() {
-		// Use a different context in case we failed because the original context failed.
-		s.setLastErrorNonFatal(s.ctx, repo, err)
-	}()
-
-	syncer, err := s.GetVCSSyncer(ctx, repo)
-	if err != nil {
-		return "", errors.Wrap(err, "get VCS syncer")
-	}
-
-	// We may be attempting to clone a private repo so we need an internal actor.
-	remoteURL, err := s.getRemoteURL(actor.WithInternalActor(ctx), repo)
-	if err != nil {
-		return "", err
-	}
-
-	// isCloneable causes a network request, so we limit the number that can
-	// run at one time. We use a separate semaphore to cloning since these
-	// checks being blocked by a few slow clones will lead to poor feedback to
-	// users. We can defer since the rest of the function does not block this
-	// goroutine.
-	ctx, cancel, err := s.acquireCloneableLimiter(ctx)
-	if err != nil {
-		return "", err // err will be a context error
-	}
-	defer cancel()
-
-	if err = s.RPSLimiter.Wait(ctx); err != nil {
-		return "", err
-	}
-
-	if err := syncer.IsCloneable(ctx, repo, remoteURL); err != nil {
-		redactedErr := urlredactor.New(remoteURL).Redact(err.Error())
-		return "", errors.Errorf("error cloning repo: repo %s not cloneable: %s", repo, redactedErr)
-	}
-
-	// Mark this repo as currently being cloned. We have to check again if someone else isn't already
-	// cloning since we released the lock. We released the lock since isCloneable is a potentially
-	// slow operation.
-	lock, ok := s.Locker.TryAcquire(dir, "starting clone")
-	if !ok {
-		// Someone else beat us to it
-		status, _ := s.Locker.Status(dir)
-		return status, nil
-	}
-
-	if s.skipCloneForTests {
-		lock.Release()
-		return "", nil
-	}
-
-	if opts.Block {
-		ctx, cancel, err := s.acquireCloneLimiter(ctx)
-		if err != nil {
-			return "", err
-		}
-		defer cancel()
-
-		// We are blocking, so use the passed in context.
-		err = s.doClone(ctx, repo, dir, syncer, lock, remoteURL, opts)
-		err = errors.Wrapf(err, "failed to clone %s", repo)
-		return "", err
-	}
-
-	// We push the cloneJob to a queue and let the producer-consumer pipeline take over from this
-	// point. See definitions of cloneJobProducer and cloneJobConsumer to understand how these jobs
-	// are processed.
-	s.CloneQueue.Push(&cloneJob{
-		repo:      repo,
-		dir:       dir,
-		syncer:    syncer,
-		lock:      lock,
-		remoteURL: remoteURL,
-		options:   opts,
-	})
-
-	return "", nil
-}
-
-func (s *Server) doClone(
-	ctx context.Context,
-	repo api.RepoName,
-	dir common.GitDir,
-	syncer vcssyncer.VCSSyncer,
-	lock RepositoryLock,
-	remoteURL *vcs.URL,
-	opts CloneOptions,
-) (err error) {
-	logger := s.Logger.Scoped("doClone").With(log.String("repo", string(repo)))
-
-	defer lock.Release()
-	defer func() {
-		if err != nil {
-			repoCloneFailedCounter.Inc()
-		}
-	}()
-	if err := s.RPSLimiter.Wait(ctx); err != nil {
-		return err
-	}
-	ctx, cancel := context.WithTimeout(ctx, conf.GitLongCommandTimeout())
-	defer cancel()
-
-	dstPath := string(dir)
-	if !opts.Overwrite {
-		// We clone to a temporary directory first, so avoid wasting resources
-		// if the directory already exists.
-		if _, err := os.Stat(dstPath); err == nil {
-			return &os.PathError{
-				Op:   "cloneRepo",
-				Path: dstPath,
-				Err:  os.ErrExist,
-			}
-		}
-	}
-
-	// We clone to a temporary location first to avoid having incomplete
-	// clones in the repo tree. This also avoids leaving behind corrupt clones
-	// if the clone is interrupted.
-	tmpDir, err := gitserverfs.TempDir(s.ReposDir, "clone-")
-	if err != nil {
-		return err
-	}
-	defer os.RemoveAll(tmpDir)
-	tmpPath := filepath.Join(tmpDir, ".git")
-
-	// It may already be cloned
-	if !repoCloned(dir) {
-		if err := s.DB.GitserverRepos().SetCloneStatus(ctx, repo, types.CloneStatusCloning, s.Hostname); err != nil {
-			s.Logger.Error("Setting clone status in DB", log.Error(err))
-		}
-	}
-	defer func() {
-		// Use a background context to ensure we still update the DB even if we time out
-		if err := s.DB.GitserverRepos().SetCloneStatus(context.Background(), repo, cloneStatus(repoCloned(dir), false), s.Hostname); err != nil {
-			s.Logger.Error("Setting clone status in DB", log.Error(err))
-		}
-	}()
-
-	logger.Info("cloning repo", log.String("tmp", tmpDir), log.String("dst", dstPath))
-
-	progressReader, progressWriter := io.Pipe()
-	// We also capture the entire output in memory for the call to SetLastOutput
-	// further down.
-	// TODO: This might require a lot of memory depending on the amount of logs
-	// produced, the ideal solution would be that readCloneProgress stores it in
-	// chunks.
-	output := &linebasedBufferedWriter{}
-	eg := readCloneProgress(s.DB, logger, lock, io.TeeReader(progressReader, output), repo)
-
-	cloneErr := syncer.Clone(ctx, repo, remoteURL, dir, tmpPath, progressWriter)
-	progressWriter.Close()
-
-	if err := eg.Wait(); err != nil {
-		s.Logger.Error("reading clone progress", log.Error(err))
-	}
-
-	// best-effort update the output of the clone
-	if err := s.DB.GitserverRepos().SetLastOutput(context.Background(), repo, output.String()); err != nil {
-		s.Logger.Error("Setting last output in DB", log.Error(err))
-	}
-
-	if cloneErr != nil {
-		// TODO: Should we really return the entire output here in an error?
-		// It could be a super big error string.
-		return errors.Wrapf(cloneErr, "clone failed. Output: %s", output.String())
-	}
-
-	if testRepoCorrupter != nil {
-		testRepoCorrupter(ctx, common.GitDir(tmpPath))
-	}
-
-	if err := postRepoFetchActions(ctx, logger, s.DB, s.Hostname, s.RecordingCommandFactory, s.ReposDir, repo, common.GitDir(tmpPath), remoteURL, syncer); err != nil {
-		return err
-	}
-
-	if opts.Overwrite {
-		// remove the current repo by putting it into our temporary directory, outside of the git repo.
-		err := fileutil.RenameAndSync(dstPath, filepath.Join(tmpDir, "old"))
-		if err != nil && !os.IsNotExist(err) {
-			return errors.Wrapf(err, "failed to remove old clone")
-		}
-	}
-
-	if err := os.MkdirAll(filepath.Dir(dstPath), os.ModePerm); err != nil {
-		return err
-	}
-	if err := fileutil.RenameAndSync(tmpPath, dstPath); err != nil {
-		return err
-	}
-
-	logger.Info("repo cloned")
-	repoClonedCounter.Inc()
-
-	s.Perforce.EnqueueChangelistMappingJob(perforce.NewChangelistMappingJob(repo, dir))
-
-	return nil
-}
-
-// linebasedBufferedWriter is an io.Writer that writes to a buffer.
-// '\r' resets the write offset to the index after last '\n' in the buffer,
-// or the beginning of the buffer if a '\n' has not been written yet.
-//
-// This exists to remove intermediate progress reports from "git clone
-// --progress".
-type linebasedBufferedWriter struct {
-	// writeOffset is the offset in buf where the next write should begin.
-	writeOffset int
-
-	// afterLastNewline is the index after the last '\n' in buf
-	// or 0 if there is no '\n' in buf.
-	afterLastNewline int
-
-	buf []byte
-}
-
-func (w *linebasedBufferedWriter) Write(p []byte) (n int, err error) {
-	l := len(p)
-	for {
-		if len(p) == 0 {
-			// If p ends in a '\r' we still want to include that in the buffer until it is overwritten.
-			break
-		}
-		idx := bytes.IndexAny(p, "\r\n")
-		if idx == -1 {
-			w.buf = append(w.buf[:w.writeOffset], p...)
-			w.writeOffset = len(w.buf)
-			break
-		}
-		w.buf = append(w.buf[:w.writeOffset], p[:idx+1]...)
-		switch p[idx] {
-		case '\n':
-			w.writeOffset = len(w.buf)
-			w.afterLastNewline = len(w.buf)
-			p = p[idx+1:]
-		case '\r':
-			// Record that our next write should overwrite the data after the most recent newline.
-			// Don't slice it off immediately here, because we want to be able to return that output
-			// until it is overwritten.
-			w.writeOffset = w.afterLastNewline
-			p = p[idx+1:]
-		default:
-			panic(fmt.Sprintf("unexpected char %q", p[idx]))
-		}
-	}
-	return l, nil
-}
-
-// String returns the contents of the buffer as a string.
-func (w *linebasedBufferedWriter) String() string {
-	return string(w.buf)
-}
-
-// Bytes returns the contents of the buffer.
-func (w *linebasedBufferedWriter) Bytes() []byte {
-	return w.buf
-}
 
 func postRepoFetchActions(
 	ctx context.Context,
@@ -1271,264 +469,7 @@ func postRepoFetchActions(
 	return errs
 }
 
-// readCloneProgress scans the reader and saves the most recent line of output
-// as the lock status, writes to a log file if siteConfig.cloneProgressLog is
-// enabled, and optionally to the database when the feature flag `clone-progress-logging`
-// is enabled.
-func readCloneProgress(db database.DB, logger log.Logger, lock RepositoryLock, pr io.Reader, repo api.RepoName) *errgroup.Group {
-	// Use a background context to ensure we still update the DB even if we
-	// time out. IE we intentionally don't take an input ctx.
-	ctx := featureflag.WithFlags(context.Background(), db.FeatureFlags())
-	enableExperimentalDBCloneProgress := featureflag.FromContext(ctx).GetBoolOr("clone-progress-logging", false)
-
-	var logFile *os.File
-
-	if conf.Get().CloneProgressLog {
-		var err error
-		logFile, err = os.CreateTemp("", "")
-		if err != nil {
-			logger.Warn("failed to create temporary clone log file", log.Error(err), log.String("repo", string(repo)))
-		} else {
-			logger.Info("logging clone output", log.String("file", logFile.Name()), log.String("repo", string(repo)))
-			defer logFile.Close()
-		}
-	}
-
-	dbWritesLimiter := rate.NewLimiter(rate.Limit(1.0), 1)
-	scan := bufio.NewScanner(pr)
-	scan.Split(scanCRLF)
-	store := db.GitserverRepos()
-
-	eg, ctx := errgroup.WithContext(ctx)
-	eg.Go(func() error {
-		for scan.Scan() {
-			progress := scan.Text()
-			lock.SetStatus(progress)
-
-			if logFile != nil {
-				// Failing to write here is non-fatal and we don't want to spam our logs if there
-				// are issues
-				_, _ = fmt.Fprintln(logFile, progress)
-			}
-			// Only write to the database persisted status if line indicates progress
-			// which is recognized by presence of a '%'. We filter these writes not to waste
-			// rate-limit tokens on log lines that would not be relevant to the user.
-			if enableExperimentalDBCloneProgress {
-				if strings.Contains(progress, "%") && dbWritesLimiter.Allow() {
-					if err := store.SetCloningProgress(ctx, repo, progress); err != nil {
-						logger.Error("error updating cloning progress in the db", log.Error(err))
-					}
-				}
-			}
-		}
-		if err := scan.Err(); err != nil {
-			return err
-		}
-
-		return nil
-	})
-
-	return eg
-}
-
-// scanCRLF is similar to bufio.ScanLines except it splits on both '\r' and '\n'
-// and it does not return tokens that contain only whitespace.
-func scanCRLF(data []byte, atEOF bool) (advance int, token []byte, err error) {
-	if atEOF && len(data) == 0 {
-		return 0, nil, nil
-	}
-	trim := func(data []byte) []byte {
-		data = bytes.TrimSpace(data)
-		if len(data) == 0 {
-			// Don't pass back a token that is all whitespace.
-			return nil
-		}
-		return data
-	}
-	if i := bytes.IndexAny(data, "\r\n"); i >= 0 {
-		// We have a full newline-terminated line.
-		return i + 1, trim(data[:i]), nil
-	}
-	// If we're at EOF, we have a final, non-terminated line. Return it.
-	if atEOF {
-		return len(data), trim(data), nil
-	}
-	// Request more data.
-	return 0, nil, nil
-}
-
-var (
-	execRunning = promauto.NewGaugeVec(prometheus.GaugeOpts{
-		Name: "src_gitserver_exec_running",
-		Help: "number of gitserver.GitCommand running concurrently.",
-	}, []string{"cmd"})
-	execDuration = promauto.NewHistogramVec(prometheus.HistogramOpts{
-		Name:    "src_gitserver_exec_duration_seconds",
-		Help:    "gitserver.GitCommand latencies in seconds.",
-		Buckets: trace.UserLatencyBuckets,
-	}, []string{"cmd", "status"})
-
-	searchRunning = promauto.NewGauge(prometheus.GaugeOpts{
-		Name: "src_gitserver_search_running",
-		Help: "number of gitserver.Search running concurrently.",
-	})
-	searchDuration = promauto.NewHistogramVec(prometheus.HistogramOpts{
-		Name:    "src_gitserver_search_duration_seconds",
-		Help:    "gitserver.Search duration in seconds.",
-		Buckets: []float64{0.01, 0.05, 0.1, 0.2, 0.5, 1, 2, 5, 10, 30},
-	}, []string{"error"})
-	searchLatency = promauto.NewHistogram(prometheus.HistogramOpts{
-		Name:    "src_gitserver_search_latency_seconds",
-		Help:    "gitserver.Search latency (time until first result is sent) in seconds.",
-		Buckets: []float64{0.01, 0.05, 0.1, 0.2, 0.5, 1, 2, 5, 10, 30},
-	})
-
-	pendingClones = promauto.NewGauge(prometheus.GaugeOpts{
-		Name: "src_gitserver_clone_queue",
-		Help: "number of repos waiting to be cloned.",
-	})
-	lsRemoteQueue = promauto.NewGauge(prometheus.GaugeOpts{
-		Name: "src_gitserver_lsremote_queue",
-		Help: "number of repos waiting to check existence on remote code host (git ls-remote).",
-	})
-	repoClonedCounter = promauto.NewCounter(prometheus.CounterOpts{
-		Name: "src_gitserver_repo_cloned",
-		Help: "number of successful git clones run",
-	})
-	repoCloneFailedCounter = promauto.NewCounter(prometheus.CounterOpts{
-		Name: "src_gitserver_repo_cloned_failed",
-		Help: "number of failed git clones",
-	})
-)
-
 var headBranchPattern = lazyregexp.New(`HEAD branch: (.+?)\n`)
-
-func (s *Server) doRepoUpdate(ctx context.Context, repo api.RepoName, revspec string) (err error) {
-	tr, ctx := trace.New(ctx, "doRepoUpdate", repo.Attr())
-	defer tr.EndWithErr(&err)
-
-	s.repoUpdateLocksMu.Lock()
-	l, ok := s.repoUpdateLocks[repo]
-	if !ok {
-		l = &locks{
-			once: new(sync.Once),
-			mu:   new(sync.Mutex),
-		}
-		s.repoUpdateLocks[repo] = l
-	}
-	once := l.once
-	mu := l.mu
-	s.repoUpdateLocksMu.Unlock()
-
-	// doBackgroundRepoUpdate can block longer than our context deadline. done will
-	// close when its done. We can return when either done is closed or our
-	// deadline has passed.
-	done := make(chan struct{})
-	err = errors.New("another operation is already in progress")
-	go func() {
-		defer close(done)
-		once.Do(func() {
-			mu.Lock() // Prevent multiple updates in parallel. It works fine, but it wastes resources.
-			defer mu.Unlock()
-
-			s.repoUpdateLocksMu.Lock()
-			l.once = new(sync.Once) // Make new requests wait for next update.
-			s.repoUpdateLocksMu.Unlock()
-
-			err = s.doBackgroundRepoUpdate(repo, revspec)
-			if err != nil {
-				// We don't want to spam our logs when the rate limiter has been set to block all
-				// updates
-				if !errors.Is(err, ratelimit.ErrBlockAll) {
-					s.Logger.Error("performing background repo update", log.Error(err))
-				}
-
-				// The repo update might have failed due to the repo being corrupt
-				var gitErr *common.GitCommandError
-				if errors.As(err, &gitErr) {
-					s.logIfCorrupt(ctx, repo, gitserverfs.RepoDirFromName(s.ReposDir, repo), gitErr.Output)
-				}
-			}
-			s.setLastErrorNonFatal(s.ctx, repo, err)
-		})
-	}()
-
-	select {
-	case <-done:
-		return errors.Wrapf(err, "repo %s:", repo)
-	case <-ctx.Done():
-		return ctx.Err()
-	}
-}
-
-var doBackgroundRepoUpdateMock func(api.RepoName) error
-
-func (s *Server) doBackgroundRepoUpdate(repo api.RepoName, revspec string) error {
-	logger := s.Logger.Scoped("backgroundRepoUpdate").With(log.String("repo", string(repo)))
-
-	if doBackgroundRepoUpdateMock != nil {
-		return doBackgroundRepoUpdateMock(repo)
-	}
-	// background context.
-	ctx, cancel1 := s.serverContext()
-	defer cancel1()
-
-	// ensure the background update doesn't hang forever
-	ctx, cancel2 := context.WithTimeout(ctx, conf.GitLongCommandTimeout())
-	defer cancel2()
-
-	// This background process should use our internal actor
-	ctx = actor.WithInternalActor(ctx)
-
-	ctx, cancel2, err := s.acquireCloneLimiter(ctx)
-	if err != nil {
-		return err
-	}
-	defer cancel2()
-
-	if err = s.RPSLimiter.Wait(ctx); err != nil {
-		return err
-	}
-
-	repo = protocol.NormalizeRepo(repo)
-	dir := gitserverfs.RepoDirFromName(s.ReposDir, repo)
-
-	remoteURL, err := s.getRemoteURL(ctx, repo)
-	if err != nil {
-		return errors.Wrap(err, "failed to determine Git remote URL")
-	}
-
-	syncer, err := s.GetVCSSyncer(ctx, repo)
-	if err != nil {
-		return errors.Wrap(err, "get VCS syncer")
-	}
-
-	// drop temporary pack files after a fetch. this function won't
-	// return until this fetch has completed or definitely-failed,
-	// either way they can't still be in use. we don't care exactly
-	// when the cleanup happens, just that it does.
-	// TODO: Should be done in janitor.
-	defer git.CleanTmpPackFiles(s.Logger, dir)
-
-	output, err := syncer.Fetch(ctx, remoteURL, repo, dir, revspec)
-	// TODO: Move the redaction also into the VCSSyncer layer here, to be in line
-	// with what clone does.
-	redactedOutput := urlredactor.New(remoteURL).Redact(string(output))
-	// best-effort update the output of the fetch
-	if err := s.DB.GitserverRepos().SetLastOutput(context.Background(), repo, redactedOutput); err != nil {
-		s.Logger.Warn("Setting last output in DB", log.Error(err))
-	}
-
-	if err != nil {
-		if output != nil {
-			return errors.Wrapf(err, "failed to fetch repo %q with output %q", repo, redactedOutput)
-		} else {
-			return errors.Wrapf(err, "failed to fetch repo %q", repo)
-		}
-	}
-
-	return postRepoFetchActions(ctx, logger, s.DB, s.Hostname, s.RecordingCommandFactory, s.ReposDir, repo, dir, remoteURL, syncer)
-}
 
 // setHEAD configures git repo defaults (such as what HEAD is) which are
 // needed for git commands to work.
